Currency system

+++ Assets/Scripts/InventoryItem.cs
@@ -3,9 +3,9 @@ public struct InventoryItem
     public string id;
     public float scaleFactor;
     public int materialIndex;
-    public float price;
+    public int price;
 
-    public InventoryItem(string id, float scaleFactor, int materialIndex, float price)
+    public InventoryItem(string id, float scaleFactor, int materialIndex, int price)
     {
         this.id = id;
         this.scaleFactor = scaleFactor;
